Add new properties, getters and setters for DAO. Add method for the CheckRepository.php. Add URL check, write to database and output of received data
Add checking request to the url for getting needed data of url to the UrlController.php.
Add creating needed data to the database in the create() method to the CheckRepository.php.
Add set needed data to the Check object to the getByUrlId() method in the CheckRepository.php.
Add new getLastStatusCode() method to the CheckRepository.php.
Add getting last status code of url checking to the getAll() method in the UrlRepository.php.
Add output last status code of urls to the urls/index.html.twig view.
Also add output status code, h1, title, description to the urls/show.html.twig view.